feat: use draggable & droppable from junipero
