+++ packages/oak/lib/core/Col/index.js
@@ -1,11 +1,10 @@
 import { useRef } from 'react';
-import { Tooltip, classNames } from '@junipero/react';
+import { Droppable, Tooltip, classNames } from '@junipero/react';
 
 import { useBuilder } from '../../hooks';
 import Catalogue from '../Catalogue';
 import Option from '../Option';
 import Element from '../Element';
-import Droppable from '../Droppable';
 import Editable from '../Editable';
 import settings from './index.settings';
 import Icon from '../Icon';
