+++ packages/oak/lib/core/Foldable/index.options.js
@@ -1,39 +1,3 @@
-import { useEffect, useRef, useState } from 'react';
-import { classNames } from '@junipero/react';
+import { DRAG_OPTION } from '../../options';
 
-import Option from '../Option';
-import Draggable from '../Draggable';
-import Text from '../Text';
-
-const DragOption = {
-  render: ({ element, elementInnerRef, className }) => {
-    const optionRef = useRef();
-    const [dragImage, setDragImage] = useState(null);
-
-    useEffect(() => {
-      setDragImage(elementInnerRef.current);
-    }, [elementInnerRef.current]);
-
-    const onDrag = () => {
-      optionRef.current?.tooltipRef?.current?.close();
-    };
-
-    return (
-      <Draggable
-        ref={optionRef}
-        onBeforeDragStart={onDrag}
-        dragImage={dragImage}
-        data={element}
-      >
-        <Option
-          onClick={e => e.preventDefault()}
-          option={{ icon: 'drag_handle' }}
-          className={classNames(className, 'oak-drag-handle')}
-          name={<Text name="core.tooltips.move" default="Move" />}
-        />
-      </Draggable>
-    );
-  },
-};
-
-export default [DragOption];
+export default [DRAG_OPTION];
